% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.tools-package.R
\docType{package}
\name{backup.tools-package}
\alias{backup.tools}
\alias{backup.tools-package}
\title{backup.tools: Tools to Create Backup Files}
\description{
Tools to create backup copies for R and RStudio setup files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GegznaV/backup.tools/}
  \item Report bugs at \url{https://github.com/GegznaV/backup.tools/issues}
}

}
\author{
\strong{Maintainer}: Vilmantas Gegzna \email{GegznaV@gmail.com} (\href{https://orcid.org/0000-0002-9500-5167}{ORCID})

}
\keyword{internal}
