% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{construct_backup_path}
\alias{construct_backup_path}
\title{Construct path to backup file.}
\usage{
construct_backup_path(
  file = NULL,
  backup_subdir = "",
  backup_stamp = get_backup_stamp()
)
}
\arguments{
\item{file}{(character) The name of file to backup.}

\item{backup_subdir}{(character) The name of subdirectory in the directory
of backups.}

\item{backup_stamp}{(character) Usually the result of \code{get_backup_stamp()}.}
}
\value{
The string with path to backup file.
}
\description{
Construct path to backup file.
}
\examples{
\dontrun{\donttest{

construct_backup_path("backup.txt")

}}
}
