% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{create_backup_copy}
\alias{create_backup_copy}
\title{Create backup file.}
\usage{
create_backup_copy(
  file = NULL,
  backup_subdir = "",
  of_what = ifelse(backup_subdir == "", "file(s)", backup_subdir),
  backup_stamp = get_backup_stamp()
)
}
\arguments{
\item{file}{(character) The name of file to backup.}

\item{backup_subdir}{(character) The name of subdirectory in the directory
of backups.}

\item{of_what}{(character) String to be used in message. Usually word "file"
or the name of \code{backup_subdir}.}

\item{backup_stamp}{(character) Usually the result of \code{get_backup_stamp()}.}
}
\description{
Create backup file.
}
