% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{get_backup_stamp}
\alias{get_backup_stamp}
\title{Get stamp for backup file names.}
\usage{
get_backup_stamp()
}
\value{
Character, e.g., \code{"__backup_200217_215436"}.
}
\description{
Get stamp for backup file names.
}
\examples{
\dontrun{\donttest{

get_backup_stamp()

}}
}
