% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{get_path_backup_dir}
\alias{get_path_backup_dir}
\alias{create_backup_dir}
\title{Path to backup directory.}
\usage{
get_path_backup_dir(...)

create_backup_dir(...)
}
\arguments{
\item{...}{The name of subtirectory. Arguments to \code{\link[fs:path]{fs::path()}}.}
}
\value{
\itemize{
\item \code{get_path_backup_dir()} returns string with path to directory.
}
}
\description{
\itemize{
\item \code{get_path_backup_dir()} gets string with path to directory for backup files.
\item \code{create_backup_dir()} creates directory for backup files.
\item 
}
}
\examples{
\dontrun{\donttest{

get_path_backup_dir()

}}
}
